requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2==2.9.5
celery[redis]==5.2.7
flower==1.2.0

